moved elements of setup_environment into different files corresponding to what they're containing. Also getting into Help system.
